Now working Restricted Hartree fock code
It is now working completely and has been tested on the He atom.
However, could do with a few minor modifications in order to read the
number of electrons and nuclear charge from the input file. Also
calculates the energy using the fock, hamiltonian and density matrices
rancher than the coulombs and exchanged integrals.